add filters query calculation
