+++ src/ui/ColumnsFilter/FilterTypeSelector/FilterTypeSelector.js
@@ -3,22 +3,9 @@ import { withTranslation } from 'react-i18next'
 import { HTMLSelect } from '@blueprintjs/core'
 import _keys from 'lodash/keys'
 
-import { propTypes, defaultProps } from './FilterTypeSelector.props'
+import FILTER_TYPES from 'var/filterTypes'
 
-const FILTER_TYPES = {
-  GT: '$gt',
-  GTE: '$gte',
-  LT: '$lt',
-  LTE: '$lte',
-  NOT: '$not',
-  LIKE: '$like',
-  EQ: '$eq',
-  NE: '$ne',
-  IN: '$in',
-  NIN: '$nin',
-  IS_NULL: '$isNull',
-  IS_NOT_NULL: '$isNotNull',
-}
+import { propTypes, defaultProps } from './FilterTypeSelector.props'
 
 class FilterTypeSelector extends React.PureComponent {
   onChange = (e) => {
@@ -33,7 +20,9 @@ class FilterTypeSelector extends React.PureComponent {
     return (
       <HTMLSelect value={value} onChange={this.onChange} className='columns-filter-item-filter'>
         <option value='' />
-        {_keys(FILTER_TYPES).map(filter => <option key={filter} value={filter}>{filter}</option>)}
+        {_keys(FILTER_TYPES).map(filter => (
+          <option key={filter} value={FILTER_TYPES[filter]}>{filter}</option>
+        ))}
       </HTMLSelect>
     )
   }
